+++ CHANGELOG.md
@@ -9,9 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `Trace` bot adaptor ([#104][pr104])
 - `HasPayload`, `Request` and `Requester` implementations for `either::Either` ([#103][pr103])
 
 [pr103]: https://github.com/teloxide/teloxide-core/pull/103
+[pr104]: https://github.com/teloxide/teloxide-core/pull/104
 
 ## 0.3.1 - 2021-07-07